+++ assets/src/scripts/components/hydrograph/time-series.spec.js
@@ -1,7 +1,7 @@
 import {
     timeSeriesSelector, hasTimeSeriesWithPoints, isVisibleSelector, yLabelSelector, titleSelector,
     descriptionSelector, currentVariableTimeSeriesSelector, allTimeSeriesSelector, tsTimeZoneSelector,
-    secondaryYLabelSelector} from './time-series';
+    getAllTimeSeriesForCurrentVariable, getAllMethodsForCurrentVariable, secondaryYLabelSelector} from './time-series';
 
 
 const TEST_DATA = {
@@ -12,6 +12,7 @@ const TEST_DATA = {
                 startTime: 1520351100000,
                 endTime: 1520948700000,
                 variable: '45807197',
+                method: 69929,
                 points: [{
                     value: 10,
                     qualifiers: ['P'],
@@ -33,7 +34,30 @@ const TEST_DATA = {
                 tsKey: 'compare:P7D',
                 startTime: 1520351100000,
                 endTime: 1520948700000,
-                variables: '45807196',
+                variable: '45807196',
+                method: 69931,
+                points: [{
+                    value: 1,
+                    qualifiers: ['P'],
+                    approved: false,
+                    estimated: false
+                }, {
+                    value: 2,
+                    qualifiers: ['P'],
+                    approved: false,
+                    estimated: false
+                }, {
+                    value: 3,
+                    qualifiers: ['P'],
+                    approved: false,
+                    estimated: false
+                }]
+            },
+            '00010:2': {tsKey: 'compare:P7D',
+                startTime: 1520351100000,
+                endTime: 1520948700000,
+                variable: '45807196',
+                method: 69930,
                 points: [{
                     value: 1,
                     qualifiers: ['P'],
@@ -55,7 +79,8 @@ const TEST_DATA = {
                 tsKey: 'compare:P7D',
                 startTime: 1520351100000,
                 endTime: 1520948700000,
-                variables: '45807195',
+                variable: '45807195',
+                method: 69929,
                 points: [{
                     value: 68,
                     qualifiers: ['P'],
@@ -78,6 +103,7 @@ const TEST_DATA = {
                 startTime: 1520351100000,
                 endTime: 1520948700000,
                 variable: '45807197',
+                method: 69929,
                 points: [{
                     value: 10,
                     qualifiers: ['P'],
@@ -132,6 +158,28 @@ const TEST_DATA = {
                 variableName: 'Temperature, water, degrees Fahrenheit',
                 variableDescription: 'Water Temperature in Fahrenheit',
                 oid: '45807195'
+            },
+            '55807196' : {
+                variableCode: {
+                    value: '11111'
+                },
+                variableName: 'My variable',
+                variableDescription: 'My awesome variable',
+                oid: '55807196'
+            }
+        },
+        methods: {
+            69329: {
+                methodDescription: '',
+                methodID: 69928
+            },
+            69330: {
+                methodDescription: '4.1 ft from riverbed (middle)',
+                methodID: 69930
+            },
+            69331: {
+                methodDescription: '1.0 ft from riverbed (bottom)',
+                methodID: 69931
             }
         },
         queryInfo: {
@@ -305,6 +353,149 @@ describe('TimeSeries module', () => {
         });
     });
 
+    describe('getAllTimeSeriesForCurrentVariable', () => {
+
+        it('Expect no time series if the current variable is not set', () => {
+            const newTestData = {
+                ...TEST_DATA,
+                timeSeriesState: {
+                }
+            };
+            expect(getAllTimeSeriesForCurrentVariable(newTestData)).toEqual({});
+        });
+
+        it('Expect no time series if the current variable does not have any timeSeries', () => {
+            const newTestData = {
+                ...TEST_DATA,
+                timeSeriesState: {
+                    ...TEST_DATA.timeSeriesState,
+                    currentVariableID: '55807196'
+                }
+            };
+            expect(getAllTimeSeriesForCurrentVariable(newTestData)).toEqual({});
+        });
+
+        it('Expect all time series for the current variable', () => {
+            const newTestData = {
+                ...TEST_DATA,
+                timeSeriesState: {
+                    ...TEST_DATA.timeSeriesState,
+                    currentVariableID: '45807196'
+                }
+            };
+            expect(getAllTimeSeriesForCurrentVariable(newTestData)).toEqual({
+                '00010': {
+                    tsKey: 'compare:P7D',
+                    startTime: 1520351100000,
+                    endTime: 1520948700000,
+                    variable: '45807196',
+                    method: 69931,
+                    points: [{
+                        value: 1,
+                        qualifiers: ['P'],
+                        approved: false,
+                        estimated: false
+                    }, {
+                        value: 2,
+                        qualifiers: ['P'],
+                        approved: false,
+                        estimated: false
+                    }, {
+                        value: 3,
+                        qualifiers: ['P'],
+                        approved: false,
+                        estimated: false
+                    }]
+                },
+                '00010:2': {
+                    tsKey: 'compare:P7D',
+                    startTime: 1520351100000,
+                    endTime: 1520948700000,
+                    variable: '45807196',
+                    method: 69930,
+                    points: [{
+                        value: 1,
+                        qualifiers: ['P'],
+                        approved: false,
+                        estimated: false
+                    }, {
+                        value: 2,
+                        qualifiers: ['P'],
+                        approved: false,
+                        estimated: false
+                    }, {
+                        value: 3,
+                        qualifiers: ['P'],
+                        approved: false,
+                        estimated: false
+                    }]
+                }
+            });
+        });
+    });
+
+    describe('getAllMethodsForCurrentVariable', () => {
+        it('Expect empty array if current variable has no time series', () => {
+            const newTestData = {
+                ...TEST_DATA,
+                timeSeriesState: {
+                    ...TEST_DATA.timeSeriesState,
+                    currentVariableID: '55807196'
+                }
+            };
+            expect(getAllMethodsForCurrentVariable(newTestData)).toEqual([]);
+        });
+
+        it('Expect method ids for current variable', () => {
+            const newTestData = {
+                ...TEST_DATA,
+                series: {
+                    ...TEST_DATA.series,
+                    timeSeries: {
+                        ...TEST_DATA.series.timeSeries,
+                        '00010:current:P30D': {
+                            tsKey: 'current:P30D',
+                            startTime: 1520351100000,
+                            endTime: 1520948700000,
+                            variable: '45807196',
+                            method: 69931,
+                            points: [{
+                                value: 1,
+                                qualifiers: ['P'],
+                                approved: false,
+                                estimated: false
+                            }, {
+                                value: 2,
+                                qualifiers: ['P'],
+                                approved: false,
+                                estimated: false
+                            }, {
+                                value: 3,
+                                qualifiers: ['P'],
+                                approved: false,
+                                estimated: false
+                            }]
+                        }
+                    }
+                },
+                timeSeriesState: {
+                    ...TEST_DATA.timeSeriesState,
+                    currentVariableID: '45807196'
+                }
+            };
+            const result = getAllMethodsForCurrentVariable(newTestData);
+            expect(result.length).toEqual(2);
+            expect(result).toContain({
+                methodDescription: '4.1 ft from riverbed (middle)',
+                methodID: 69930
+            });
+            expect(result).toContain({
+                methodDescription: '1.0 ft from riverbed (bottom)',
+                methodID: 69931
+            });
+        });
+    });
+
     describe('timeSeriesSelector', () => {
 
         it('should return the selected time series', () => {
@@ -329,7 +520,8 @@ describe('TimeSeries module', () => {
                         approved: false,
                         estimated: false
                     }],
-                    variable: '45807197'
+                    variable: '45807197',
+                    method: 69929
                 }
             });
             expect(timeSeriesSelector('current','P30D')(TEST_DATA)).toEqual({
@@ -338,6 +530,7 @@ describe('TimeSeries module', () => {
                     startTime: 1520351100000,
                     endTime: 1520948700000,
                     variable: '45807197',
+                    method: 69929,
                     points: [{
                         value: 10,
                         qualifiers: ['P'],