+++ assets/src/scripts/components/hydrograph/tooltip.js
@@ -11,7 +11,7 @@ import { classesForPoint, MASK_DESC } from './drawing-data';
 import { layoutSelector } from './layout';
 import { xScaleSelector, yScaleSelector } from './scales';
 import { tsTimeZoneSelector, TEMPERATURE_PARAMETERS } from './time-series';
-import { getCurrentVariable, getCurrentParmCd } from '../../selectors/time-series-selector';
+import { getCurrentVariable, getCurrentParmCd, getCurrentMethodID } from '../../selectors/time-series-selector';
 import config from '../../config';
 import { mediaQuery, convertCelsiusToFahrenheit, convertFahrenheitToCelsius } from '../../utils';
 
@@ -114,7 +114,7 @@ const unitCodeSelector = createSelector(
     variable => variable ? variable.unit.unitCode : null
 );
 
-const createTooltipTextGroup = function (elem, {currentPoints, comparePoints, qualifiers, unitCode, ianaTimeZone, layout, currentParmCd}, textGroup) {
+const createTooltipTextGroup = function (elem, {currentPoints, comparePoints, currentMethodID, qualifiers, unitCode, ianaTimeZone, layout, currentParmCd}, textGroup) {
     // Find the width of the between the y-axis and margin and set the tooltip margin based on that number
     const adjustMarginOfTooltips = function (elem) {
         // set a base number of pixels to bump the tooltips away from y-axis and compensate for slight under reporting
@@ -123,6 +123,20 @@ const createTooltipTextGroup = function (elem, {currentPoints, comparePoints, qu
         let marginAdjustment = layout.margin.left + baseMarginOffsetTextGroup;
         elem.style('margin-left', marginAdjustment + 'px');
     };
+    const currentPointsWithMethod = Object.keys(currentPoints).map((tsKey) => {
+        const methodID = tsKey.split(':')[0];
+        return {
+            ...currentPoints[tsKey],
+            methodID: methodID
+        };
+    });
+    const comparePointsWithMethod = Object.keys(comparePoints).map((tsKey) => {
+        const methodID = tsKey.split(':')[0];
+        return {
+            ...comparePoints[tsKey],
+            methodID: methodID
+        };
+    });
 
     // Put the circles in a container so we can keep the their position in the
     // DOM before rect.overlay, to prevent the circles from receiving mouse
@@ -133,7 +147,7 @@ const createTooltipTextGroup = function (elem, {currentPoints, comparePoints, qu
             .call(adjustMarginOfTooltips);
     }
 
-    const data = Object.values(currentPoints).concat(Object.values(comparePoints));
+    const data = Object.values(currentPointsWithMethod).concat(Object.values(comparePointsWithMethod));
     const texts = textGroup
         .selectAll('div')
         .data(data);
@@ -177,7 +191,6 @@ const createTooltipTextGroup = function (elem, {currentPoints, comparePoints, qu
     // Update the text and backgrounds of all tooltip labels
     const merge = texts.merge(newTexts)
         .interrupt()
-        .style('opacity', '1')
         .call(adjustTooltipFontSize);
 
     merge
@@ -188,6 +201,7 @@ const createTooltipTextGroup = function (elem, {currentPoints, comparePoints, qu
             text.attr('class', d => `${d.tsKey}-tooltip-text`);
             text.classed('approved', classes.approved);
             text.classed('estimated', classes.estimated);
+            text.classed('not-current-method', currentMethodID !== parseInt(datum.methodID));
         });
 
     return textGroup;
@@ -202,6 +216,7 @@ export const createTooltipText = function (elem) {
     elem.call(link(createTooltipTextGroup, createStructuredSelector({
         currentPoints: tsCursorPointsSelector('current'),
         comparePoints: tsCursorPointsSelector('compare'),
+        currentMethodID: getCurrentMethodID,
         qualifiers: qualifiersSelector,
         unitCode: unitCodeSelector,
         layout: layoutSelector,