+++ assets/src/scripts/selectors/time-series-selector.js
@@ -8,13 +8,15 @@ import { createSelector } from 'reselect';
  */
 export const getVariables = state => state.series.variables ? state.series.variables : null;
 
-export const getMethods = state => state.series.methods ? state.series.methods : null;
+export const getMethods = state => state.series.methods ? state.series.methods : {};
 
 export const getQueryInfo = state => state.series.queryInfo || {};
 
 export const getRequests = state => state.series.requests || {};
 
 export const getCurrentVariableID = state => state.timeSeriesState.currentVariableID;
+
+export const getCurrentMethodID = state => state.timeSeriesState.currentMethodID;
 
 export const getCurrentDateRange = state => state.timeSeriesState.currentDateRange;
 
@@ -114,7 +116,7 @@ export const getTSRequest = memoize((tsKey, period, parmCd) => createSelector(
     (requests, tsRequestKey) => requests[tsRequestKey] || {}
 ));
 
-/*
+ /*
  * @param {String} tsKey - current or compare
  * @param {String} or null period - date range of interest specified as an ISO-8601 duration. If null, P7D is assumed
  * @param {String} or null parmCd - Only need to specify if period is something other than P7D or null