+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased](https://github.com/usgs/waterdataui/compare/waterdataui-0.19.0...master)
 ### Added
 - Added a second Y-axis for temperature parameters
+- Added picker to pick a method description that will be highlighted in the hydrograph.
 
 ### Changed
 - Changed how the lower bound on symlog scales are calculated to be closer to the minimum y-value on the plot